+++ packages/example/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @grid-table/example
 
+## 0.0.46
+
+### Patch Changes
+
+- Updated dependencies
+  - @grid-table/view@0.6.33
+  - @grid-table/core@0.6.33
+  - @grid-table/excel@0.5.36
+  - @grid-table/pivot@0.5.39
+  - @grid-table/basic@0.6.33
+
 ## 0.0.45
 
 ### Patch Changes